# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SSBUTools API",
        "description": "RESTful API providing data related to Super Smash Bros. Ultimate assets",
        "contact": {
            "name": "Raynald Mirville",
            "email": "admin@ssbutools.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/v1",
            "description": "SSBUTools API Server"
        }
    ],
    "paths": {
...
        "/stages/piece-maps": {
            "get": {
                "summary": "Returns an index of stage piece maps",
                "description": "Returns an index of stage piece maps",
                "operationId": "pieceMapsIndex",
                "responses": {
                    "200": {
                        "description": "An index of all stage piece maps",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "_links": {
                                            "$ref": "#/components/schemas/links_index"
                                        },
                                        "_embedded": {
                                            "properties": {
                                                "pieceMaps": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/stage_piece_map_summary"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/resource_not_found"
                    },
                    "500": {
                        "$ref": "#/components/responses/server_error"
                    }
                }
            }
        },
        "/stages/piece-maps/{piece-map-name}": {
            "get": {
                "summary": "Returns a stage piece map",
                "description": "Returns a mapping of stages to selected pieces",
                "operationId": "getPieceMapByName",
                "parameters": [
                    {
                        "name": "piece-map-name",
                        "in": "path",
                        "description": "Name of stage piece map",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A stage piece map",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/stage_piece_map"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/resource_not_found"
                    },
                    "500": {
                        "$ref": "#/components/responses/server_error"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "link": {
                "properties": {
                    "href": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "links": {
                "properties": {
                    "self": {
                        "$ref": "#/components/schemas/link"
                    }
                },
                "type": "object"
            },
            "links_index": {
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/links"
                    },
                    {
                        "properties": {
                            "index": {
                                "$ref": "#/components/schemas/link"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
...
            "error": {
                "properties": {
                    "status": {
                        "type": "number"
                    },
                    "type": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "error_response": {
                "properties": {
                    "_links": {
                        "properties": {
                            "home": {
                                "$ref": "#/components/schemas/link"
                            }
                        },
                        "type": "object"
                    },
                    "_error": {
                        "$ref": "#/components/schemas/error"
                    }
                },
                "type": "object"
            },
            "id": {
                "type": "string"
            },
...
            "stage_piece_map_item": {
                "properties": {
                    "lvd": {
                        "type": "string"
                    },
                    "pieceName": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "stage_piece_map": {
                "properties": {
                    "_links": {
                        "$ref": "#/components/schemas/links_index"
                    },
                    "id": {
                        "$ref": "#/components/schemas/id"
                    },
                    "maps": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/stage_piece_map_item"
                        }
                    }
                },
                "type": "object"
            },
            "stage_piece_map_summary": {
                "properties": {
                    "_links": {
                        "$ref": "#/components/schemas/links"
                    },
                    "id": {
                        "$ref": "#/components/schemas/id"
                    }
                },
                "type": "object"
            }
        },
        "responses": {
...
            "resource_not_found": {
                "description": "Requested resource not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/error_response"
                        }
                    }
                }
            },
            "server_error": {
                "description": "Unexpected error occured",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/error_response"
                        }
                    }
                }
            }
        }
    },
...
    "security": [
        []
    ]
}